backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
google-generativeai
pydantic
python-dotenv
joblib
matplotlib
seaborn
